requirements.txt
opencv-python-headless
streamlit
torch
sahi
torchvision
pillow
yolov8
ultralytics
pyrebase4
numpy
pandas

